Move ro.dalvik.vm.native.bridge to /system/build.prop
ro.dalvik.vm.native.bridge does not actually need to be defined
in default.prop. This allows removing it from the boot image,
and therefore patching it using the flashable ZIP.

This also allows simplifying the make files, because target/houdini.mk
does no longer need to be included unless Houdini should be bundled
with the system. Everything else is handled in the board file.